added req validation for all apis

+++ Controllers/admin.controller.js
@@ -38,7 +38,13 @@ async function handleGetMyCars (req, res) {
 async function handleDeleteCar(req, res) {
     try{
         let {id} = req.body;
+        if (!id) {
+            return res.status(400).send({ message: "Car ID is required", error: true });
+        }
         let data = await carDetailModel.deleteOne({_id:id});
+        if (data.deletedCount === 0) {
+            return res.status(404).send({ message: "Car not found", error: true });
+        }
         res.status(200).send({ message: "Car is deleted", error: false });
     }
     catch(err) {
@@ -50,7 +56,17 @@ async function handleDeleteCar(req, res) {
 async function handleUpdateCar(req, res) {
     try{
         let {id,charge} = req.body;
+
+        if (!id || charge === undefined) {
+            return res.status(400).send({ message: "Car ID and charge are required", error: true });
+        }
+        
         let data = await carDetailModel.updateOne({_id:id},{$set:{charge}});
+
+        if (data.nModified === 0) {
+            return res.status(404).send({ message: "Car not found or no update made", error: true });
+        }
+
         res.status(200).send({ message: "Car is updated", error: false });
     }
     catch(err) {
